+++ docs/changelog.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Added:***
+
+- Support `dict`
+- Support `typing.TypedDict`
+
 ## 0.1.0 - 2024-09-18
 
 This is the initial public release.